+++ lessweb/__init__.py
@@ -1,3 +1,3 @@
-from .bridge import (Bridge, autowire, delete_mapping, get_mapping,
+from .bridge import (Bridge, autowire, config, delete_mapping, get_mapping,
                      patch_mapping, post_mapping, put_mapping, rest_error,
                      rest_mapping, rest_response, service)